Move server API code into the server module to keep it away from the client code
