+++ .github/workflows/generate-api-client.yml
@@ -1,0 +1,68 @@
+name: Generate API Client
+
+on:
+  push:
+    branches:
+      - develop
+    paths:
+      - 'observatory-api/observatory/api/server/openapi.yaml.jinja2'
+
+jobs:
+  generate:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Install OpenAPI Generator
+        run: |
+          mkdir -p ~/bin/openapitools
+          curl https://raw.githubusercontent.com/OpenAPITools/openapi-generator/master/bin/utils/openapi-generator-cli.sh > ~/bin/openapitools/openapi-generator-cli
+          chmod u+x ~/bin/openapitools/openapi-generator-cli
+          export PATH=$PATH:~/bin/openapitools/
+          openapi-generator-cli version
+
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.7
+
+      - name: Install packages
+        run: |
+          python -m pip install --upgrade pip
+          pip install observatory-api
+
+      - name: Generate OpenAPI specification
+        run: |
+          observatory-api generate ...
+
+      - name: Generate Observatory API Client
+        run |
+        openapi-generator-cli generate -i observatory-api/openapi.yaml -g python -c observatory-api/api-config.yaml -o observatory-api
+
+      - name: Create Pull Request
+        id: cpr
+        uses: peter-evans/create-pull-request@v3
+        with:
+          token: ${{ secrets.PAT }}
+          commit-message: Update Observatory API Client
+          committer: GitHub <noreply@github.com>
+          author: ${{ github.actor }} <${{ github.actor }}@users.noreply.github.com>
+          signoff: false
+          branch: feature-api-client
+          delete-branch: true
+          title: 'Update Observatory API Client'
+          body: |
+            Update Observatory API Client
+            - Generated a new Observatory API client based on the latest openapi.yaml.jinja2 template.
+            - Auto-generated by [create-pull-request][1]
+
+            [1]: https://github.com/peter-evans/create-pull-request
+          labels: |
+            api
+            automated pr
+          assignees: jdddog
+          reviewers: jdddog
+          team-reviewers: |
+            data-science-team
+          draft: false
